Consolidate gRPC integration tests into fewer key files
Remove redundant test files (test_bulk_large_upload, test_bulk_stream,
test_bulk_standard_client, test_bulk_request, test_bulk_request_integration,
test_client_operations) and merge unique tests into test_bulk.py.

Remaining files focus on correctness:
- test_bulk.py: index, create, update, delete, mixed ops, response format, status codes
- test_exceptions.py: error handling
- test_fallback.py: REST fallback verification
- test_grpc_transport.py: transport-level tests
